+++ plugins/tweak/tweak.cpp
@@ -14,6 +14,7 @@
 #include "modules/Materials.h"
 #include "modules/MapCache.h"
 #include "modules/Buildings.h"
+#include "modules/Filesystem.h"
 
 #include "MiscUtils.h"
 
@@ -96,6 +97,7 @@
 #include "tweaks/nestbox-color.h"
 #include "tweaks/shift-8-scroll.h"
 #include "tweaks/stable-cursor.h"
+#include "tweaks/title-start-rename.h"
 #include "tweaks/tradereq-pet-gender.h"
 
 using std::set;
@@ -218,6 +220,8 @@ DFhackCExport command_result plugin_init (color_ostream &out, std::vector <Plugi
         "  tweak shift-8-scroll [disable]\n"
         "    Gives Shift+8 (or *) priority when scrolling menus, instead of \n"
         "    scrolling the map\n"
+        "  tweak title-start-rename [disable]\n"
+        "    Adds a safe rename option to the title screen \"Start Playing\" menu\n"
         "  tweak tradereq-pet-gender [disable]\n"
         "    Displays the gender of pets in the trade request list\n"
 //        "  tweak military-training [disable]\n"
@@ -276,6 +280,9 @@ DFhackCExport command_result plugin_init (color_ostream &out, std::vector <Plugi
     TWEAK_HOOK("shift-8-scroll", shift_8_scroll_hook, feed);
 
     TWEAK_HOOK("stable-cursor", stable_cursor_hook, feed);
+
+    TWEAK_HOOK("title-start-rename", title_start_rename_hook, feed);
+    TWEAK_HOOK("title-start-rename", title_start_rename_hook, render);
 
     TWEAK_HOOK("tradereq-pet-gender", pet_gender_hook, render);